requirements.txt
Flask
flask-login
flask-mail
flask-mongoengine
mongoengine
python-dotenv
Werkzeug
itsdangerous
pymongo
dnspython
email-validator
Pillow
python-dateutil
requests
blinker
click
Jinja2
MarkupSafe
six
urllib3
